15. April 2016
-Namesteno da se moze slati sa logike preko 1200 mm
-Ide u kocku, ne preskace korake
-Veliki robot apdejtovan PID

+++ Headers/funkcije.h
@@ -14,8 +14,9 @@ void nuliraj_poziciju_robota(void);
 void zadaj_X_Y_teta(signed long x, signed long y, signed long teta_des, unsigned char dir);
 void zadaj_X_Y(signed long x, signed long y, unsigned char dir);
 void zadaj_teta(signed long teta_des, unsigned char dir);
-void idi_pravo(unsigned int x, unsigned int y, unsigned int ugao);
-void idi_unazad( unsigned int x,  unsigned int y,  unsigned int ugao);
+void idi_pravo(signed long x, signed long y, signed long ugao);
+void idi_unazad(signed long x, signed long y, signed long ugao);
+
 
 void sendMsg(char *poruka);
 void SendChar(char c);
